% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{firedam}
\alias{firedam}
\title{FIREDAM data set}
\format{
## `firedam`
A data set with distance and damgage
}
\usage{
firedam
}
\description{
A data set that describes the distance and damage done
}
\keyword{datasets}
