% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MATH4753FALLmeye0060.R
\docType{package}
\name{MATH4753FALLmeye0060}
\alias{MATH4753FALLmeye0060}
\title{MATH4753FALLmeye0060 Package}
\description{
Contains a function for reading CSV files easily
}
\author{
Nathan Meyer \email{nathan.n.meyer-1@ou.edu}
}
