% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{myboot2 function}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{this is the amount that you want to iterate}

\item{x}{is the sample of particular distribution}

\item{fun}{function to be applied}

\item{alpha}{1 - the confidence level}

\item{cx}{line weight}

\item{...}{extra params}
}
\description{
myboot2 function
}
