% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclt.R
\name{myclt}
\alias{myclt}
\title{My CLT function}
\usage{
myclt(n, iter, a = 0, b = 5)
}
\arguments{
\item{n}{sample size}

\item{iter}{how many iterations}

\item{a}{lower bound}

\item{b}{upper bound}
}
\value{
a vector of the sum of the random uniform distribution
}
\description{
My CLT function
}
