% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myhyper.R
\name{myhyper}
\alias{myhyper}
\title{Creates a table for hypergeometric simulation with bargraphs for results}
\usage{
myhyper(iter = 100, N = 20, r = 12, n = 5)
}
\arguments{
\item{iter}{Number of times of iteration}

\item{N}{Sample size}

\item{r}{Number of desired amount}

\item{n}{amount to choose from}
}
\value{
bargraphs and table
}
\description{
Creates a table for hypergeometric simulation with bargraphs for results
}
