% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Creates a plot with a curve plotting the probabilities with a given interval a}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{Average for a normal curve}

\item{sigma}{Standard deviation of a normal curve}

\item{a}{P(Y <= a)}
}
\value{
plot and list
}
\description{
Creates a plot with a curve plotting the probabilities with a given interval a
}
