% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myread.R
\name{myread}
\alias{myread}
\title{Allows R to read in a CSV file in the current directory}
\usage{
myread(csv)
}
\arguments{
\item{csv}{CSV file}
}
\value{
data of CSV file
}
\description{
Insert the CSV file as a function argument, make sure CSV file is in the current working directory
}
