% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Overbooking tickets calculator}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{N variable for total seats available}

\item{gamma}{Probability the plane will be overbooked}

\item{p}{Probability of a 'show'}
}
\value{
Makes two plots for discrete and continous cases, prints lists of n tickets to be sold for both cases
}
\description{
Overbooking tickets calculator
}
